Collect GitHub models in the same location
