+++ src/main/kotlin/no/risc/github/GithubHelper.kt
@@ -3,118 +3,10 @@
 package no.risc.github
 
 import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.JsonIgnoreUnknownKeys
-import no.risc.utils.KOffsetDateTimeSerializer
+import no.risc.github.models.GithubCreateNewBranchPayload
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
 import java.time.OffsetDateTime
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class GithubReferenceObjectDTO(
-    val ref: String,
-    val url: String,
-    @SerialName("object")
-    val shaObject: GithubRefShaDTO,
-)
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class GithubRefShaCommitCommiter(
-    @Serializable(KOffsetDateTimeSerializer::class)
-    @SerialName("date")
-    val dateTime: OffsetDateTime,
-)
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class GithubRefShaCommit(
-    val committer: GithubRefShaCommitCommiter,
-)
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class GithubRefShaDTO(
-    val sha: String,
-    val url: String,
-)
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class GithubRefCommitDTO(
-    val sha: String,
-    val url: String,
-    val commit: GithubRefShaCommit,
-)
-
-@Serializable
-data class GithubCreateNewBranchPayload(
-    val nameOfNewBranch: String,
-    val shaOfLatestDefault: String,
-) {
-    fun toContentBody(): String = "{ \"ref\":\"$nameOfNewBranch\", \"sha\": \"$shaOfLatestDefault\" }"
-}
-
-@Serializable
-data class GithubCreateNewPullRequestPayload(
-    val title: String,
-    val body: String,
-    val repositoryOwner: String,
-    val branch: String,
-    val baseBranch: String,
-) {
-    fun toContentBody(): String =
-        "{ \"title\":\"$title\", \"body\": \"$body\", \"head\": \"$repositoryOwner:$branch\", \"base\": \"$baseBranch\" }"
-}
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class GithubPullRequestObject(
-    @SerialName("html_url")
-    val url: String,
-    val title: String,
-    @SerialName("created_at")
-    @Serializable(KOffsetDateTimeSerializer::class)
-    val createdAt: OffsetDateTime,
-    val head: GithubPullRequestHead,
-    val base: GithubPullRequestHead,
-    val number: Int,
-    val user: GitHubPullRequestUser,
-)
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class GitHubPullRequestUser(
-    val login: String,
-)
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class GithubCommitObject(
-    val commit: Commit,
-)
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class Commit(
-    val message: String,
-    val committer: Committer,
-)
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class Committer(
-    val date: String,
-    val name: String,
-)
-
-@Serializable
-@JsonIgnoreUnknownKeys
-data class GithubPullRequestHead(
-    val ref: String,
-)
 
 @Component
 class GithubHelper(
